requirements.txt
customtkinter==5.2.1
google-generativeai==0.3.2
Pillow==10.2.0
reportlab==4.1.0
matplotlib==3.8.2
bcrypt==4.1.2
python-dotenv==1.0.1
packaging==23.2 